fix(lint): split large files
